watchlist/requirements.txt
alembic==1.7.1
amqp-worker==0.2
attrs==21.2.0
blinker==1.4
certifi==2021.5.30
Chameleon==3.9.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
debut==0.9.9
diff-match-patch==20120106
dnspython==2.1.0
docutils==0.17.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Reuploaded==1.1.0
Flask-Script==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
hupper==1.10.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
linecache2==1.0.0
lockfile==0.12.2
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.2
passlib==1.7.4
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==3.2.1
pathtools==0.1.2
pika==1.2.0
plaster==1.0
plaster-pastedeploy==0.7
popular==0.1.5
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyramid==2.0
pyramid-addons==0.21
pyramid-chameleon==0.3
pyramid-exclog==1.0
pyramid-layout==1.0
pyramid-mailer==0.15.1
pyramid-tm==2.4
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-ldap==3.3.1
PyYAML==5.4.1
repoze.sendmail==4.4.1
requests==2.26.0
review==1510
six==1.16.0
sqla-mixins==0.6
SQLAlchemy==1.4.23
traceback2==1.4.0
transaction==3.0.1
translationstring==1.4
unittest2==1.1.0
urllib3==1.26.6
venusian==3.0.0
visitor==0.1.3
WebOb==1.8.7
Werkzeug==2.0.1
WTForms==2.3.3
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
